예시

+++ Lambda/src/section05/SupplierExample2.java
@@ -3,7 +3,7 @@
  * - 리턴값이 있는 메소드를 가지고 있는 함수적 인터페이스
  * @FunctionalInterface
  * public interface IntSupplier {
- *    int getAsInt();
+ *     int getAsInt();
  * }
  */
 package section05;
@@ -13,15 +13,14 @@ import java.util.function.Supplier;
 public class SupplierExample2 {
 
 	public static void main(String[] args) {
-		Supplier<Integer> supplier = () -> {
-			int num = (int)(Math.random() * 6 + 1);
+		Supplier<Long> supplier = () -> {
+			long num = (long)(Math.random() * 6 + 1);
 			return num;
 		};
-		
+
 		for(int x = 0; x < 6; x++) {
-			int num = supplier.get();
+			long num = supplier.get();
 			System.out.println("눈의 수: " + num);
 		}
 	}
-
 }
